styles and refactor
