+++ pandemania/src/components/CountryIcon/CountryIcon.jsx
@@ -22,7 +22,7 @@ export default CountryIcon;
 
 CountryIcon.propTypes = {
 	code: propTypes.string,
-	size: propTypes.number,
+	size: propTypes.oneOf([16, 24, 32, 48, 64]),
 	name: propTypes.string,
 	theme: propTypes.string,
 };
