+++ pandemania/src/components/Layout/Layout.css
@@ -5,8 +5,19 @@ h1 {
 .title{
   font-size: 1rem;
   margin: 0 0rem 0.5rem;
+  display: flex;
+  align-items: center;
+}
+
+.title span{
+  margin-left: 0.5rem;
 }
 
 .bold{
   font-weight: bold;
+}
+
+img{
+  margin-left: 0.25rem;
+  margin-top: auto;
 }
