+++ pandemania/src/components/Layout/Layout.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import './Layout.css';
-import AreaDateTimeCountryData from '../graphs/AreaDateTimeCountryData/AreaDateTimeCountryData';
+import AreaDateTimeCountryData from '../ChartWrapper/ChartWrapper';
 import CountryPicker from '../CountryPicker/CountryPicker';
+import CountryIcon from '../CountryIcon/CountryIcon';
 
 const Layout = () => {
 	const [pickedCountry, setPickedCountry] = useState({
@@ -12,12 +13,13 @@ const Layout = () => {
 	return (
 		<>
 			<header>
-				<h1>PandeMania</h1>
+				<h1>PandeMania 🦠</h1>
 			</header>
 			<main>
 				<p className="title">
 					Cases and recoveries for whole country:
 					{pickedCountry && <span className="bold"> {pickedCountry.country}</span>}
+					<CountryIcon code={pickedCountry.ISO2} size={24} name={pickedCountry.country} />
 				</p>
 				<CountryPicker setPickedCountry={(value) => setPickedCountry(value)} />
 				<AreaDateTimeCountryData countryName={pickedCountry.slug} />
